Relocate file closer to it's relatives
