+++ src/behaviours/helm-charts/remove-helm-repository-from-list-of-active-repository-in-preferences.test.ts
@@ -11,7 +11,7 @@ import asyncFn from "@async-fn/jest";
 import execFileInjectable from "../../common/fs/exec-file.injectable";
 import helmBinaryPathInjectable from "../../main/helm/helm-binary-path.injectable";
 import getActiveHelmRepositoriesInjectable from "../../main/helm/repositories/get-active-helm-repositories/get-active-helm-repositories.injectable";
-import type { HelmRepo } from "../../common/helm-repo";
+import type { HelmRepo } from "../../common/helm/helm-repo";
 import callForPublicHelmRepositoriesInjectable from "../../renderer/components/+preferences/kubernetes/helm-charts/adding-of-public-helm-repository/public-helm-repositories/call-for-public-helm-repositories.injectable";
 import type { AsyncResult } from "../../common/utils/async-result";
