+++ src/common/helm/add-helm-repository-channel.injectable.ts
@@ -3,7 +3,7 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 import { getInjectable } from "@ogre-tools/injectable";
-import type { HelmRepo } from "../helm-repo";
+import type { HelmRepo } from "./helm-repo";
 import type { RequestChannel } from "../utils/channel/request-channel-injection-token";
 import { requestChannelInjectionToken } from "../utils/channel/request-channel-injection-token";
 import type { AsyncResult } from "../utils/async-result";
