+++ src/main/helm/__tests__/helm-service.test.ts
@@ -7,7 +7,7 @@ import { getDiForUnitTesting } from "../../getDiForUnitTesting";
 import listHelmChartsInjectable from "../helm-service/list-helm-charts.injectable";
 import getActiveHelmRepositoriesInjectable from "../repositories/get-active-helm-repositories/get-active-helm-repositories.injectable";
 import type { AsyncResult } from "../../../common/utils/async-result";
-import type { HelmRepo } from "../../../common/helm-repo";
+import type { HelmRepo } from "../../../common/helm/helm-repo";
 
 jest.mock("../helm-chart-manager");
