+++ src/renderer/components/+preferences/kubernetes/helm-charts/adding-of-custom-helm-repository/submit-custom-helm-repository.injectable.ts
@@ -3,7 +3,7 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 import { getInjectable } from "@ogre-tools/injectable";
-import type { HelmRepo } from "../../../../../../common/helm-repo";
+import type { HelmRepo } from "../../../../../../common/helm/helm-repo";
 import addHelmRepositoryInjectable from "../adding-of-public-helm-repository/select-helm-repository/add-helm-repository.injectable";
 import hideDialogForAddingCustomHelmRepositoryInjectable from "./dialog-visibility/hide-dialog-for-adding-custom-helm-repository.injectable";
