+++ src/renderer/components/+preferences/kubernetes/helm-charts/adding-of-public-helm-repository/select-helm-repository/select-helm-repository.injectable.ts
@@ -5,7 +5,7 @@
 import { getInjectable } from "@ogre-tools/injectable";
 import addHelmRepositoryInjectable from "./add-helm-repository.injectable";
 import type { SelectOption } from "../../../../../select";
-import type { HelmRepo } from "../../../../../../../common/helm-repo";
+import type { HelmRepo } from "../../../../../../../common/helm/helm-repo";
 import type { SingleValue } from "react-select";
 import removeHelmRepositoryInjectable from "../../remove-helm-repository.injectable";
